+++ src/common/metadata-utils/index.ts
@@ -1,10 +1,11 @@
 import { SfdxError } from "@salesforce/core";
 import * as c from "chalk";
+import * as changedGitFiles from "changed-git-files";
 import * as extractZip from "extract-zip";
 import * as fs from "fs-extra";
 import * as path from "path";
 import * as sortArray from "sort-array";
-import { elapseEnd, elapseStart, execCommand, execSfdxJson, filterPackageXml, uxLog } from "../../common/utils";
+import { elapseEnd, elapseStart, execCommand, execSfdxJson, filterPackageXml, isGitRepo, uxLog } from "../../common/utils";
 import { CONSTANTS } from "../../config";
 import { PACKAGE_ROOT_DIR } from "../../settings";
 import { getCache, setCache } from "../cache";
@@ -667,6 +668,30 @@ Issue tracking: https://github.com/forcedotcom/cli/issues/2426`),
       }),
     });
     return metadataResp.value;
+  }
+
+  // List updated files and reformat them as string
+  public static async listChangedFiles(): Promise<string[]> {
+    if (!isGitRepo()) {
+      return [];
+    }
+    const files = await new Promise<string[]>((resolve) => {
+      changedGitFiles((err: any, result: any[]) => {
+        if (result == null) {
+          console.warn(JSON.stringify(err, null, 2));
+          resolve([]);
+        }
+        resolve(result);
+      });
+    });
+    const filesTextLines = files
+      .sort((a: any, b: any) => (a.filename > b.filename ? 1 : -1))
+      .map((fileInfo: any) => {
+        fileInfo.shortFileName = fileInfo.filename.replace("force-app/main/default/", "");
+        return fileInfo;
+      })
+      .map((fileInfo: any) => `${fileInfo.shortFileName} - ${fileInfo.status}`);
+    return filesTextLines;
   }
 }